Start on colour handling
Define types for RGBA and HSLA colours, and conversions between them.

Experimenting with a highly typeful style of programming for the numbers
stored with the colours. Not sure if this will work within the beginner
context; we'll see.

Need to add some unit tests for the conversion. Just grabbed the
algorithms from random places and there is no guarantee they work or
that I have implemented them correctly.