Refactor platform specific rendering
This commits separates the backend API for rendering graphics from the
process of rendering an Image. This allows more code reuse and provides
a better foundation for the Doodle case study.

backend.Canvas defines an imperative API for rendering to a 2D vector
graphics context. It's API generally matches that found in JS and Java
2D, and provides a common abstraction uniting the two platforms.
Platform specific implementations live in js and jvm

backend.Interpreter defines a (simplified) typeclass for rendering
images to a Canvas. There is currently a single implementation in
backend.StandardInterpreter.

syntax.draw adds an enrichment to Images to draw an image using an
implicit Canvas and Interpreter. This allows us to write

`Circle(20).draw`

vs

`draw(Circle(20))`