Rename Drawable#draw to #image
Image is a better name, as the method returns an image. The name draw
also conflicted with the implicit draw method that renders an Image.

As part of this change I removed some JVM code that was no longer
needed: DoodleFrame and DoodlePanel, and a Main class that was not used
as far as I'm aware.