Give greater control over drawing size and origin
Canvas adds a new method setSize, which sets the size of the canvas.
This removes the complex code in the Java 2D canvas for recalculating
the size, which should save some time,which should save some time, and
the canvas size is easily calculated from the Image bounding box.

The draw methods in the interpreter allow the origin and size to be
inferred from the Image (the default) or specified as parameters. The
later is necessary for rendering animations. When rendering animations
we don't want to continually change the origin or canvas size every
frame as the animation will not render as expected.