Prevent angles being created with NaN
Pointed out by @ilya-epifanov. Thanks!

The constructor of Angle converts NaN to 0 radians, which avoids some errors in color conversion with gray colors. It's something of an arbitrary decision to do this and not throw an error, but it works for this use case and doesn't seem harmful in general.

Starting adding ScalaTest / ScalaCheck tests to the project, removing old tests in the JS subproject that are replaced by these new tests.