Shift Random into an object-as-module
Like a true Typelevel hipster.

Definining the Random type in the package object was having some strange
interactions with the Random object in the REPL---use the Random object
once and it would suddenly disappear from the REPL.

This style of programming makes more sense for the way Random is
written. The object acts like a package.