Major refactoring of `Chart` internal representation
This is NOT the final `Chart` representation, but one sufficient to continue
exploration of the API. In this representation we differentiate at the type AND
value level between numerical and categorical data. This is needed so we can
dispatch to appropriate interpreters at runtime. The alternative, building an
interpreter at compile time, I deemed too heavy weight.

This representation has limited extensibility, so will eventually be replaced. I
believe that a "recursion schemes" style representation will allow the
extensibility we want.

The graphs don't render as well as they might. In particular, the spacing around
the tick marks and axis labels is poor. I would ideally enhance the `Image`
language to support more layout options to make this easier.

Finally, added some basic support for styling. This is just a sketch. The final
result will be much more expressive.