Allow direct drawing on the Canvas
`Image.draw` allows the user to provide a method that draws directly on the
canvas. This is generally faster than constructing an `Image` graph representing
an equivalent output and is the only way to render images with millions of
components. It is also arguably more suitable for certain types of output (e.g
images where there is little structure.)

Closes #43