tweak dependencies
the cats part is needed now for Doodle to be included in the Scala
community build.  depending only on the parts of cats that are actually
needed is also generally good practice.

the ScalaCheck version bump isn't strictly necessary at the moment
but will help ensure community build compatibility moving forward.

the ScalaTest version bump is just for good measure.