Cleanly shutdown Java2D engine and default to it in the console
The `stop` method on the Java2D engine disposes of all frames, which in turn
kills daemon threads and allows the JVM to exit.

Since the Java2D engine can be stopped and started multiple times, unlike the
JavaFX engine which can only be started and stopped once per JVM invocation,
default to it in the console.

Closes #1