Further work to support animation
Use a ScheduledThreadPoolExecutor for the frame timer, which helps avoid
deadlocking the Swing Event Dispatch Thread (which is used for the Swing Timer)

Don't store all the intermediate images in the `Java2DFrame`. This wasn't doing
anything useful and is a memory leak.

Automatically dispose of windows when they're closed. This avoids Swing
preventing the JVM from shutting down, which in turn makes the `console` more
usable. Previously in some situations you couldn't exit the `console` and had to
kill the JVM.

Avoid NPE in initialising `Java2DPanel`

Syntax for animation

Miscellaneous cleanups and UX improvements.