Proof-of-concept explorer implementation
An explorer automatically derives a UI to explore an function's inputs, based on
their type, and renders the output. This makes for a much better creative coding
workflow, as modifying parameters is far faster without an edit-compile cycle.

This is a proof-of-concept implementation. Code is a bit messy and there are not
many explorer UI types defined, but it works. See
`doodle/explore/Example.scala`.