Steps towards affine transforms
Affine transformations are implemented, but buggy. Layout is correct in simple cases but origns are not transformed correctly in some cases (e.g. rotations of groups of primitive images).

Generic algebra produces a reified output. This means effects are only executed once, as they should be.

scalafmt'ed the code, which led to many trivial changes. Oops. Should have done this in a separate PR.