Abstract over definition of Frame
The old `Frame` type doesn't work for SVG, where the user must specify
an DOM element to bind and some of the options don't make sense. So
abstract over `Frame` by making it a type parameter in all the relevant
places. Required introducing a `DefaultRenderer` so the user can still
draw without specifying a Frame and gets a sensible result.