Generic algebras abstract over the final effect type
In the previous design the effect output by the generic algebras was
fixed: they always produced a list of reified commands. This is not
expressive enough to add interactivity. For example, if we want
a (sub-)Image to support mouse events we need to know the nesting
structure  at the time we're rendering so we can add an event handler to
just that image.

In the new design the effect type can be whatever is best for the
backend. For example the SVG backend could work directly with SVG
elements, and by doing so attach event handlers and so on.

Currently all backends still use the reified representation. The reified
algebras are now a particular implementation of the generic algebras and
live in `doodle.algebra.generic.reified`.