Update tests but hit initialization order issue
Tests are updated to match changes to generic algebras. However there is
currently an initialization order issue in the `applyF` instance in
`GenericLayout`. I either get a runtime exception due to initialization
order or, if I change all `val` to `def` I get a stack overflow from
repeated calls to summon the instance. Creating a brach to experiment
with a fix.