First working SVG animation
Animation works on SVG. See `Example.scala`

This is not a wholly satisfactory solution:

1. The call to `animate` doesn't actually run the animation. The user
must later call `unsafeRunAsync`.
2. The `redraw` method on `Canvas`, which should allow redraw when the browser
is ready to redraw, doesn't work.
3. The animation is a bit janky.

To fix 1 requires some changes to the effect result types. I was
assuming I could call `IO` synchronously, so going `IO[A] => A`. This is
not the case on JS. I cheesed this by returning `IO[A]`. A better
solution is to change the result type of `draw` and `animate` to
`Unit` (and allow an optional callback)

Fixing 2 is probably just better knowledge of the `Observable` API.

Fixing 3 I'm not sure on. It might just be that rendering SVG nodes on
every frame is too slow. Canvas might be a better target for animation,
but we lose interactivity (mouse events etc.) or perhaps using a DOM
diffing library to minimise DOM changes is the solution.

This commit includes refactoring `SvgRenderer` and `Canvas` so that
`Canvas` does most of the drawing. This allows us to reduce the
rerendering to the child nodes of the <svg> tag.