Normalize effect syntax
Animation and rendering syntax now consistently run the
effect (animation used to return `IO[A]`), run effects asychronously,
and return `Unit`.

Previously rendering syntax ran the effect _synchronously_. The same
could not be done for animations on the JS runtime. Thus there was an
inconsistency between the two types of syntax.

Also added more methods to handle situations where a `Canvas` is
available, a `Frame` is available, or some sensible default so the user
doesn't need to specify anything for ease of use. Normalized naming
conventions of these methods.