Consistent use of `Picture` as value type
Algebras return `F[A]`, with the expectation that we write methods with
signatures like

```scala
def someMethod[Alg[x] <: SomeAlgebra[x], F[_]](implicit alg: Alg[F]):
F[Unit] = ???
```

This is very tedious and easy to get wrong. Also, a method is not a
value so composition is limited. The `Picture` makes it much
simpler for the end user to write code using algebras _and_ it creates a
value that can be passed around, composed, etc.

Syntax methods were defined for `Picture` and `F[A]`. For a binary
operation there were methods for `Picture op Picture` and `F[A] op F[A]`
but not other possibilities. There is also ambiguity between the various
syntax as the `Picture` and `F[A]` type often end up looking the same
due to various aliases and constraints. This all meant the programming
modeo could be difficult to understand. Sometimes you'd end up working
with a `F[A]` and sometimes a `Picture` and you'd have to do
conversions, and sometimes the compiler would just give up.

In this change the algebras still work with `F[A]` but the syntax
consistently converts to `Picture` and *only* `Picture`. This removes
the issues above, whilst the unwrapped operations are still available
should the extra allocation be a concern.