Add Reactor project
Reactors provide a simple way to create animations. Currently they only work
with the Image library, and do not support input. In the future we might extend
them to the core algebras and support input methods that the `interact` project
provides.

Inspired by the library of the same name in Pyret.