Implement gradients in SVG
Add support for gradients to the SVG backend. I'm not convinced that
gradients are actually working correctly but this is sufficient to get
something out.

To share more code between the JS and JVM backends I went full ML-module
style (Cake pattern) on the implementation. It was an interesting
experiment. I don't like the resulting code as it is very hard to
reason about. There is also a weird interaction with package objects
that requires code duplication. It would be interesting to reimplmement
this using type parameters and values rather than type members and
mixins. I think this would be easier to reason about, though perhaps
cumbersom to write.

- Add some examples of gradients in svg and image
- `Writer` is contravariant
- Update to Scala.js 0.6.29
- Remove dead `SvgGraphicsContext`