Initial version of easy to use animation utilities
The Transducer type is a utility for animations that can be represented
as finite state machines (plus a mapping from state to output, which
makes it a transducer). Transducers have a lot of algebraic structure,
which provides a lot of utilities for free (and make it easy to compose
transducers).

Enumerator makes it easy to express simple ranges like (0.upto(100)) and
turn these into the basis of animations.