Add `ToPicture` algebra and syntax
The `ToPicture` algebra provides a generic way to convert some type into
a `Picture`. The syntax adds a method `toPicture` so one can call, e.g.

```scala
someBase64Value.toPicture[Algebra, Drawing]
```

To facilitate this change there is a new wrapper type,
`doodle.core.Base64`, which indicates base 64 encoded data and stores
the format as a type parameter. The `Base64` algebra has been updated
accordingly.

There are implementations of `ToPicture` for the Java2d backend for `Base64` and
`BufferedImage`.

This replaces some of the functionality of the `Bitmap` algebra and it
is likely that the `Bitmap` algebra will be removed or reworked in the future.