Move some of the examples into their own project.
The SVG examples export a class into the JS top level, which pollutes
the user's name space. Shift them into a separate project that isn't
part of the core distribution to avoid this. It also gives more scope to
develop examples that are, for example, cross back-end or otherwise show
more advanced features.