POC improving constructor type inference
This is an attempt to improve type inference for constructors in
algebras. See #115. The plan is to get rid of the type parameters that are
currently need for constructors if only targeting a single backend.
E.g., instead of writing

  `circle[Algebra, Drawing](10)`

you could just write

  `Picture.circle(10)`

There is less noise here, and IDE method completion when using
the `Picture` object.