Remove F / Drawing type as a parameter for Picture
This type becomes a type member instead of being a type parameter. This
means Pictures are no longer parameterized by their output effect type
and it thus becomes easier to create pictures that are cross platform,
and hence build libraries that are cross platform.

A large change but I think it's a good one.