draw() on the JVM blocks until drawing completes
Swing uses a daemon thread. The JVM will exit is only daemon threads
exist. This means that simple programs that draw and immediately exit
will usually not display any output unless the main thread waits for the
drawing to complete. This is a problem for many students using IDEs.

This change:

1. Changes how `draw` is implemented, so it blocks on the JVM but does
not block on JS (JS cannot block.)

2. Refactors render syntax to be more consistent and implements derived
methods in terms of a few primitives.

3. Refactors JVM `Frame` to follow a more consistent builder style, as
used in other parts of Doodle (e.g. `Reactor`)

4. Documents `Frame` and drawing

5. Updates examples etc. as necessary