Implement ToPicture for ImageBitmap
I expect this will fail at runtime as I believe ImageBitmap is not an
instance of HTMLElement. Fixing this will require changes to Scala JS
DOM to add overloads to drawImage. There is not currently a way
to (easily) test this, as we don't have methods to extract an
ImageBitmap from a Drawing.

Refactor code to avoid duplication in implementation.

Format and fix.